+++ gui/gui_menu.cpp
@@ -23,16 +23,6 @@ void Gui::win_menu() {
             
             ImGui::Text("scene: %s", scene_list.at(active_scene).name.c_str());
             
-            if (Graphics::camera.proj_type & Graphics::PROJECTION_SCALED)
-                ImGui::Text("res: %dx%d (x%d)", Config::resolution.x, Config::resolution.y, Graphics::win.scale);
-            else
-                ImGui::Text("res: %dx%d", Graphics::win.size.x, Graphics::win.size.y);
-            
-            if (Graphics::camera.proj_type & Graphics::PROJECTION_PERSPECTIVE)
-                ImGui::Text("proj: perspective");
-            if (Graphics::camera.proj_type & Graphics::PROJECTION_ORTHOGRAPHIC)
-                ImGui::Text("proj: orthographic");
-            
             ImGui::EndMenu();
         }
         
@@ -45,7 +35,45 @@ void Gui::win_menu() {
             ImGui::EndMenu();
         }
         
-        //---Options---
+        //---Camera---
+        if (ImGui::BeginMenu("camera")) {
+            if (Graphics::camera.proj_type & Graphics::PROJECTION_SCALED)
+                ImGui::Text("res: %dx%d (x%d)", Config::resolution.x, Config::resolution.y, Graphics::win.scale);
+            else
+                ImGui::Text("res: %dx%d", Graphics::win.size.x, Graphics::win.size.y);
+            
+            if (Graphics::camera.proj_type & Graphics::PROJECTION_PERSPECTIVE)
+                ImGui::Text("proj: perspective");
+            if (Graphics::camera.proj_type & Graphics::PROJECTION_ORTHOGRAPHIC)
+                ImGui::Text("proj: orthographic");
+            
+            ImGui::AlignTextToFramePadding();
+            ImGui::Text("pos:");
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            slider("", Graphics::camera.pos.x, 1.0f, -10000.0f, 10000.0f);
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            slider("", Graphics::camera.pos.y, 1.0f, -10000.0f, 10000.0f);
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            slider("", Graphics::camera.pos.z, 1.0f, -10000.0f, 10000.0f);
+            
+            auto normalize_direction = [&](){
+                Graphics::camera.direction = glm::normalize(Graphics::camera.direction);
+            };
+            
+            ImGui::AlignTextToFramePadding();
+            ImGui::Text("dir:");
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            if (slider("", Graphics::camera.direction.x, 0.01f, -1.0f, 1.0f)) normalize_direction();
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            if (slider("", Graphics::camera.direction.y, 0.01f, -1.0f, 1.0f)) normalize_direction();
+            ImGui::SameLine(); ImGui::SetNextItemWidth(50.0f);
+            if (slider("", Graphics::camera.direction.z, 0.01f, -1.0f, 1.0f)) normalize_direction();
+            
+            ImGui::EndMenu();
+        }
+        
+        
+        //---Utils---
         if (ImGui::BeginMenu("utils")) {
             for (auto &win : windows) {
                 if (win.name == "menu") continue;